Virtual_wallet/requirements.txt
asgiref==3.10.0
colorama==0.4.6
coverage==7.10.7
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
iniconfig==2.1.0
model-bakery==1.20.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
whitenoise==6.11.0
